Abstracted stripe payments logic to lib
