+++ test/api/v3/integration/payments/stripe/GET-payments_stripe_subscribe_cancel.test.js
@@ -1,11 +1,13 @@
 import {
   generateUser,
+  generateGroup,
   translate as t,
 } from '../../../../../helpers/api-integration/v3';
+import stripePayments from '../../../../../../website/server/libs/stripePayments';
 
 describe('payments - stripe - #subscribeCancel', () => {
   let endpoint = '/stripe/subscribe/cancel';
-  let user;
+  let user, group, stripeCancelSubscriptionStub;
 
   beforeEach(async () => {
     user = await generateUser();
@@ -16,6 +18,61 @@ describe('payments - stripe - #subscribeCancel', () => {
       code: 401,
       error: 'NotAuthorized',
       message: t('missingSubscription'),
+    });
+  });
+
+  describe('success', () => {
+    beforeEach(async () => {
+      stripeCancelSubscriptionStub = sinon.stub(stripePayments, 'cancelSubscription').returnsPromise().resolves({});
+    });
+
+    afterEach(() => {
+      stripePayments.cancelSubscription.restore();
+    });
+
+    it('cancels a user subscription', async () => {
+      user = await generateUser({
+        'profile.name': 'sender',
+        'purchased.plan.customerId': 'customer-id',
+        'purchased.plan.planId': 'basic_3mo',
+        'purchased.plan.lastBillingDate': new Date(),
+        balance: 2,
+      });
+
+      await user.get(`${endpoint}?redirect=none`);
+
+      expect(stripeCancelSubscriptionStub).to.be.calledOnce;
+      expect(stripeCancelSubscriptionStub.args[0][0].user._id).to.eql(user._id);
+      expect(stripeCancelSubscriptionStub.args[0][0].groupId).to.eql(undefined);
+      expect(stripeCancelSubscriptionStub.args[0][0].headers['x-api-key']).to.eql(user.apiToken);
+      expect(stripeCancelSubscriptionStub.args[0][0].headers['x-api-user']).to.eql(user._id);
+    });
+
+    it('cancels a group subscription', async () => {
+      user = await generateUser({
+        'profile.name': 'sender',
+        'purchased.plan.customerId': 'customer-id',
+        'purchased.plan.planId': 'basic_3mo',
+        'purchased.plan.lastBillingDate': new Date(),
+        balance: 2,
+      });
+
+      group = await generateGroup(user, {
+        name: 'test group',
+        type: 'guild',
+        privacy: 'public',
+        'purchased.plan.customerId': 'customer-id',
+        'purchased.plan.planId': 'basic_3mo',
+        'purchased.plan.lastBillingDate': new Date(),
+      });
+
+      await user.get(`${endpoint}?groupId=${group._id}&redirect=none`);
+
+      expect(stripeCancelSubscriptionStub).to.be.calledOnce;
+      expect(stripeCancelSubscriptionStub.args[0][0].user._id).to.eql(user._id);
+      expect(stripeCancelSubscriptionStub.args[0][0].groupId).to.eql(group._id);
+      expect(stripeCancelSubscriptionStub.args[0][0].headers['x-api-key']).to.eql(user.apiToken);
+      expect(stripeCancelSubscriptionStub.args[0][0].headers['x-api-user']).to.eql(user._id);
     });
   });
 });
