+++ website/server/libs/payments.js
@@ -19,14 +19,10 @@ import {
 } from './errors';
 import slack from './slack';
 import nconf from 'nconf';
-import stripeModule from 'stripe';
 import {
   BadRequest,
 } from './errors';
 import cc from 'coupon-code';
-
-const stripe = stripeModule(nconf.get('STRIPE_API_KEY'));
-
 
 let api = {};
 
@@ -380,108 +376,6 @@ api.buyGems = async function buyGems (data) {
   }
 
   await data.user.save();
-};
-
-/**
- * Allows for purchasing a user subscription, group subscription or gems with Stripe
- *
- * @param  options
- * @param  options.token  The stripe token generated on the front end
- * @param  options.user  The user object who is purchasing
- * @param  options.gift  The gift details if any
- * @param  options.sub  The subscription data to purchase
- * @param  options.groupId  The id of the group purchasing a subscription
- * @param  options.email  The email enter by the user on the Stripe form
- * @param  options.headers  The request headers to store on analytics
- * @return undefined
- */
-api.payWithStripe = async function payWithStripe (options, stripeInc) {
-  let {
-    token,
-    user,
-    gift,
-    sub,
-    groupId,
-    email,
-    headers,
-    coupon,
-  } = options;
-  let response;
-  let subscriptionId;
-  // @TODO: We need to mock this, but curently we don't have correct Dependency Injection
-  let stripeApi = stripe;
-
-  if (stripeInc) stripeApi = stripeInc;
-
-  if (!token) throw new BadRequest('Missing req.body.id');
-
-  if (sub) {
-    if (sub.discount) {
-      if (!coupon) throw new BadRequest(shared.i18n.t('couponCodeRequired'));
-      coupon = await Coupon.findOne({_id: cc.validate(coupon), event: sub.key}).exec();
-      if (!coupon) throw new BadRequest(shared.i18n.t('invalidCoupon'));
-    }
-
-    let customerObject = {
-      email,
-      metadata: { uuid: user._id },
-      card: token,
-      plan: sub.key,
-    };
-
-    if (groupId) {
-      customerObject.quantity = sub.quantity;
-    }
-
-    response = await stripeApi.customers.create(customerObject);
-
-    if (groupId) subscriptionId = response.subscriptions.data[0].id;
-  } else {
-    let amount = 500; // $5
-
-    if (gift) {
-      if (gift.type === 'subscription') {
-        amount = `${shared.content.subscriptionBlocks[gift.subscription.key].price * 100}`;
-      } else {
-        amount = `${gift.gems.amount / 4 * 100}`;
-      }
-    }
-
-    response = await stripe.charges.create({
-      amount,
-      currency: 'usd',
-      card: token,
-    });
-  }
-
-  if (sub) {
-    await this.createSubscription({
-      user,
-      customerId: response.id,
-      paymentMethod: 'Stripe',
-      sub,
-      headers,
-      groupId,
-      subscriptionId,
-    });
-  } else {
-    let method = 'buyGems';
-    let data = {
-      user,
-      customerId: response.id,
-      paymentMethod: 'Stripe',
-      gift,
-    };
-
-    if (gift) {
-      let member = await User.findById(gift.uuid).exec();
-      gift.member = member;
-      if (gift.type === 'subscription') method = 'createSubscription';
-      data.paymentMethod = 'Gift';
-    }
-
-    await this[method](data);
-  }
 };
 
 module.exports = api;
